+++ .gitignore
@@ -2,3 +2,4 @@
 env
 data
 models
+__pycache__
