+++ master.py
@@ -1,163 +1,9 @@
+from src.data_preprocess import load_data
+from src.trader import FreeLaborTrader
+from src.state import State
 import pandas as pd
-import numpy as np
-import os
-import pickle
-import tensorflow as tf
-import random
 
 from tqdm import tqdm
-from collections import deque
-
-
-class FreeLaborTrader:
-    def __init__(self, state_size, action_space=4, window_size=60):
-        self.state_size = state_size
-        self.action_space = action_space
-        self.window_size = window_size
-        self.memory = deque(maxlen=2000)
-
-        self.gamma = 0.95
-        self.epsilon = 1.0
-        self.epsilon_final = 0.01
-        self.epsilon_decay = 0.995
-
-        self.model = self.build_model()
-
-    def build_model(self):
-        model = tf.keras.Sequential()
-        # TODO: Use tf.keras.layers.CuDNNLSTM
-        model.add(
-            tf.keras.layers.Dense(
-                units=32, activation="relu", input_dim=self.state_size
-            )
-        )
-        model.add(tf.keras.layers.Dense(units=128, activation="relu"))
-        model.add(tf.keras.layers.Dense(units=self.action_space, activation="linear"))
-        model.compile(loss="mean_squared_error", optimizer=tf.keras.optimizers.Adamax())
-
-        return model
-
-    def trade(self, state):
-        if random.random() <= self.epsilon:
-            return random.randrange(self.action_space)
-
-        actions = self.model.predict(state)
-        return np.argmax(actions[0])
-
-    def batch_train(self, batch_size):
-        batch = [
-            self.memory[i]
-            for i in range(len(self.memory) - batch_size + 1, len(self.memory))
-        ]
-
-        for state, action, reward, next_state, done in batch:
-            # Reward if agent is in terminal state
-            reward = reward
-
-            # TODO: Do we need diminished returns?
-            if not done:
-                reward = reward + self.gamma * np.amax(
-                    self.model.predict(next_state)[0]
-                )
-
-            target = self.model.predict(state)
-            target[0][action] = reward
-
-            self.model.fit(state, target, epochs=1, verbose=0)
-
-        if self.epsilon > self.epsilon_final:
-            self.epsilon *= self.epsilon_decay
-
-
-def load_data(file_path: str):
-    assert os.path.exists(file_path), f"{file_path} does not exist."
-    data = pd.read_csv(file_path)
-
-    # Check for the presence of specific columns
-    # TODO: DateTime mandatory if data processed by using datetime
-    required_columns = ["Open", "High", "Low", "Close", "Volume"]
-    assert set(required_columns).issubset(
-        data.columns
-    ), f"File {file_path} does not contain the required columns: {required_columns}"
-
-    return data
-
-
-class State:
-    def __init__(
-        self,
-        open: float,
-        high: float,
-        low: float,
-        close: float,
-        volume: int,
-        balance=10000.00,
-        entry_price=0.00,
-        contracts=0,
-    ):
-        self.open = open
-        self.high = high
-        self.low = low
-        self.close = close
-        self.volume = volume
-        self.balance = balance
-        self.entry_price = entry_price
-        self.contracts = contracts
-
-    def enter_long(self, entry_price: float, contracts: int, price_per_contract: float):
-        assert not self.has_position(), "Exit current position first."
-        self.entry_price = entry_price
-        self.balance -= contracts * price_per_contract
-        self.contracts = contracts
-
-    def enter_short(
-        self, entry_price: float, contracts: int, price_per_contract: float
-    ):
-        assert not self.has_position(), "Exit current position first."
-        self.entry_price = entry_price
-        self.balance += contracts * price_per_contract
-        self.contracts = contracts
-
-    def exit_position(self, exit_price: float, price_per_contract: float) -> float:
-        assert self.has_position(), "No position to exit."
-        profit = (exit_price - self.entry_price) * self.contracts * price_per_contract
-        self.balance += profit
-        self.entry_price = 0
-        self.contracts = 0
-
-        return profit
-
-    def has_position(self):
-        """
-        return: 1 for long position, -1 for short position, 0 for no position
-        """
-        if self.contracts > 0:
-            return 1
-        elif self.contracts < 0:
-            return -1
-        else:
-            return 0
-
-    def rep_position(self):
-        if not self.has_position():
-            return "No position."
-
-        position = self.has_position()
-        return f"{'Long' if position == 1 else 'Short'} position: {self.contracts} contracts entered at {self.entry_price}."
-
-    def to_df(self) -> pd.DataFrame:
-        return pd.DataFrame(
-            {
-                "open": [self.open],
-                "high": [self.high],
-                "low": [self.low],
-                "close": [self.close],
-                "volume": [self.volume],
-                "balance": [self.balance],
-                "entry_price": [self.entry_price],
-                "contracts": [self.contracts],
-            }
-        )
 
 
 def state_creator(data: pd.DataFrame, timestep: int, state: State = None):
